Setup seperate views for Candidates and recruiters
