+++ src/components/App.js
@@ -48,7 +48,7 @@ const App = () => {
         handleSignOut={(loggedIn) => setLoggedIn(loggedIn)}
       />
       {
-        (loggedIn.logged_in) ? (<Jobs />) : (<Home />)
+        (loggedIn.logged_in) ? (<Jobs user={loggedIn.user} />) : (<Home />)
       }
       <Footer />
       {console.log('from App return', loggedIn)}
