+++ src/containers/RecruiterJobs.js
@@ -1,0 +1,7 @@
+import '../styles/CandidateJobs.css';
+
+const RecruiterJobs = () => (
+  <h3 className="view">Recruiter&apos;s View</h3>
+);
+
+export default RecruiterJobs;
