+++ src/styles/CandidateJobs.css
@@ -1,0 +1,7 @@
+.view {
+  margin: 70px 140px;
+}
+
+.jobs-container {
+  margin-top: 100px;
+}
